The Flask Mega-Tutorial, Part V: User Logins
We add login support using OpenID
* Use Flask-Login and Flask-OpenID, NOTE: official github
mitsuhiko/Flask-OpenID, which depends on python-openid, which only
supports python2, in order to use in python3, using
DennisGlindhart/flask-openid fork, which depends on python3-openid,
which supports python3.
* Original tutorial has this code in app/models.py::User class:
	class User(db.Model):
		def get_id(self):
			return unicode(self.id)
NOTE the function `unicode` is in python2, for converting id to
unicode type. As in python3, function `unicode` becomes `str`.

Signed-off-by: Wang <wangwangwar@gmail.com>